W3W-23 Wallet cell background view fix
