+++ iosApp/iosApp/Common/Views/GradientView.swift
@@ -1,3 +1,7 @@
+// Created by web3d3v on 08/09/2022.
+// Copyright (c) 2022 Sons Of Crypto.
+// SPDX-License-Identifier: MIT
+
 // Created by web3d3v on 11/02/2022.
 // Copyright (c) 2022 Sons Of Crypto.
 // SPDX-License-Identifier: MIT
@@ -8,39 +12,14 @@ import UIKit
 /// stuble shadows without having to add assets to the app. `UIView` wrapper
 /// is handy as it avoids having to deal with more cumbersome `CALayer` layout
 @IBDesignable
-final class GradientView: UIView {
-    
+class GradientView: UIView {
+
     enum Direction {
         case vertical
         case horizontal
         case custom(CGPoint, CGPoint)
     }
-    
-    init() {
-        super.init(frame: .zero)
-        configureUI()
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureUI()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configureUI()
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        configureUI()
-    }
-    
-    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
-        super.traitCollectionDidChange(previousTraitCollection)
-        configureUI()
-    }
-    
+
     /// Defining the color of each gradient * stop. Defaults to nil. Animatable.
     var colors: [UIColor] {
         get {
@@ -51,15 +30,7 @@ final class GradientView: UIView {
             (layer as? CAGradientLayer)?.colors = newValue.map { $0.cgColor }
         }
     }
-    
-    var isDashboard: Bool = false {
-        
-        didSet {
-            
-            configureUI()
-        }
-    }
-    
+
     /// Defines direction of gradient in `CAGradientLayer`
     var direction: Direction = .horizontal {
         didSet {
@@ -76,51 +47,17 @@ final class GradientView: UIView {
             }
         }
     }
-    
+
     override class var layerClass: AnyClass {
         return CAGradientLayer.self
     }
-    
+
     /// two black colors with aplha at 0 and 0.25
     class func defaultShadowColors() -> [UIColor] {
         return [UIColor.black.withAlphaComponent(0), UIColor.black.withAlphaComponent(0.25)]
     }
-    
+
     override func prepareForInterfaceBuilder() {
         super.prepareForInterfaceBuilder()
-    }
-}
-
-private extension GradientView {
-    
-    func configureUI() {
-        
-        switch Theme.type {
-            
-        case .themeMiami:
-
-            let topColor = Theme.colour.gradientTop
-//            let middleTopColor = isDarkMode ? UIColor(rgb: 0xDA3E9B) : UIColor(rgb: 0x4361BC).withAlpha(0.75)
-//            let middleColor = isDarkMode ? UIColor(rgb: 0xC939AA) : UIColor(rgb: 0xBE6CF4).withAlpha(0.75)
-//            let middleBottomColor = isDarkMode ? UIColor(rgb: 0xAC36B0) : UIColor(rgb: 0xFC78A9).withAlpha(0.71)
-            let bottomColor = Theme.colour.gradientBottom
-            
-            if isDashboard {
-                colors = [
-                    topColor, bottomColor
-                ]
-                //(layer as? CAGradientLayer)?.locations = [0,0.3,0.375,0.5,0.75,1]
-            } else {
-                colors = [
-                    topColor, bottomColor
-                ]
-                //(layer as? CAGradientLayer)?.locations = [0,0.3,0.5,0.75,1]
-            }
-            
-        case .themeVanilla:
-            colors = [
-                Theme.colour.gradientTop, Theme.colour.gradientBottom
-            ]
-        }
     }
 }
