+++ iosApp/iosApp/Common/Views/ThemeGradientView.swift
@@ -1,0 +1,28 @@
+// Created by web3d3v on 11/02/2022.
+// Copyright (c) 2022 Sons Of Crypto.
+// SPDX-License-Identifier: MIT
+
+import UIKit
+
+@IBDesignable
+final class ThemeGradientView: GradientView {
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureUI()
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configureUI()
+    }
+
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+        configureUI()
+    }
+
+    private func configureUI() {
+        colors = [Theme.colour.gradientTop, Theme.colour.gradientBottom]
+    }
+}
