+++ iosApp/iosApp/Modules/Account/Account.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="HQv-1q-hcp">
             <objects>
                 <viewController storyboardIdentifier="AccountViewController" id="tlF-Di-doq" userLabel="AccountViewController" customClass="AccountViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="dSa-E0-Gsz" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="dSa-E0-Gsz" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
