+++ iosApp/iosApp/Modules/Degen/Cult/CultProposals/CultProposals.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="Z9F-7H-yvM">
             <objects>
                 <viewController storyboardIdentifier="CultProposalsViewController" title="iosApp" id="OMG-Ce-Kj4" userLabel="CultProposalsViewController" customClass="CultProposalsViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="Xez-Ps-xQi" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="Xez-Ps-xQi" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
