+++ iosApp/iosApp/Modules/KeyStore/KeyStore.storyboard
@@ -14,7 +14,7 @@
         <scene sceneID="0RO-Y9-en0">
             <objects>
                 <viewController storyboardIdentifier="KeyStoreViewController" title="iosApp" id="11H-kE-Hg7" customClass="KeyStoreViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="KvM-fo-Kai" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="KvM-fo-Kai" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
