+++ iosApp/iosApp/Modules/MnemonicImport/MnemonicImport.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="kAc-Pk-Fax">
             <objects>
                 <viewController storyboardIdentifier="MnemonicImportViewController" title="Mnemonic" id="ozv-nd-HZD" customClass="MnemonicImportViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="sR2-sg-55A" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="sR2-sg-55A" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
