+++ iosApp/iosApp/Modules/NFTSend/NFTSend.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="x5M-du-NTm">
             <objects>
                 <viewController storyboardIdentifier="NFTSendViewController" id="VSz-Bo-FCC" customClass="NFTSendViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="JGr-ZJ-egI" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="JGr-ZJ-egI" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
