+++ iosApp/iosApp/Modules/NetworkPicker/NetworkPickerView.swift
@@ -73,7 +73,7 @@ private extension NetworkPickerViewController {
         
     func configureUI() {
         
-        (view as? GradientView)?.colors = [
+        (view as? ThemeGradientView)?.colors = [
             Theme.colour.backgroundBaseSecondary,
             Theme.colour.backgroundBasePrimary
         ]
