+++ iosApp/iosApp/Modules/Settings/Settings.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="ykF-8C-r7e">
             <objects>
                 <viewController storyboardIdentifier="SettingsViewController" title="Settings" id="mov-jZ-4Mh" customClass="SettingsViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="eOT-W3-QgJ" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="eOT-W3-QgJ" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
