+++ iosApp/iosApp/Modules/TokenSwap/TokenSwap.storyboard
@@ -13,7 +13,7 @@
         <scene sceneID="xDI-Vd-biU">
             <objects>
                 <viewController storyboardIdentifier="TokenSwapViewController" id="ncT-FB-SOz" customClass="TokenSwapViewController" customModule="iosApp" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="96v-2a-j6q" customClass="GradientView" customModule="iosApp" customModuleProvider="target">
+                    <view key="view" contentMode="scaleToFill" id="96v-2a-j6q" customClass="ThemeGradientView" customModule="iosApp" customModuleProvider="target">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
